+++ src/batch/update.ts
@@ -1,12 +1,12 @@
 import { OriWriteBatch, OriDocumentReference, WriteBatchUpdate } from '../types'
 import { flatten } from '../utils'
 
-export const updateCreator = ((writeBatch: OriWriteBatch) =>
-	(reference: OriDocumentReference, data: Record<string, unknown>) => {
+export const updateCreator = (writeBatch: OriWriteBatch) =>
+	((reference: OriDocumentReference, data: Record<string, unknown>) => {
 		const ref = writeBatch.update(
 			reference,
 			// @ts-expect-error
 			flatten(data)
 		)
 		return ref
-	}) as unknown as (writeBatch: OriWriteBatch) => WriteBatchUpdate
+	}) as WriteBatchUpdate