+++ src/transaction/get.ts
@@ -1,11 +1,6 @@
 import { OriTransaction, Get, OriDocumentReference } from '../types'
 
-export const getCreator =
-	(transaction: OriTransaction): Get =>
-	// @ts-expect-error
-	async reference => {
-		const ref = await transaction.get(
-			reference as unknown as OriDocumentReference
-		)
-		return ref
-	}
+export const getCreator = (transaction: OriTransaction) =>
+	((reference: OriDocumentReference) => {
+		return transaction.get(reference) as unknown
+	}) as Get