+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # HISTORY
 
+## v1.5.0 19-July-2022
+
+- rename setting `allFieldsPossiblyUndefined` as `allFieldsPossiblyReadAsUndefined`
+- remove onSnapshot onCompletion parameter
+
 ## v1.3.0 6-May-2022
 
 - orderBy('\_name') cursor now only accept full doc path, if input type is string, require const assertion or else display "Please use const assertion" error message.